+++ embassy-usb-dfu/src/dfu.rs
@@ -183,7 +183,7 @@ impl<'d, DFU: NorFlash, STATE: NorFlash, RST: Reset, const BLOCK_SIZE: usize> Ha
             Ok(Request::GetStatus) => {
                 match self.state {
                     State::DlSync => self.state = State::Download,
-                    State::ManifestSync => self.reset.sys_reset(),
+                    State::ManifestSync => self.state = State::ManifestWaitReset,
                     _ => {}
                 }
 
@@ -200,6 +200,12 @@ impl<'d, DFU: NorFlash, STATE: NorFlash, RST: Reset, const BLOCK_SIZE: usize> Ha
                 Some(InResponse::Rejected)
             }
             _ => None,
+        }
+    }
+
+    fn reset(&mut self) {
+        if matches!(self.state, State::ManifestSync | State::ManifestWaitReset) {
+            self.reset.sys_reset()
         }
     }
 }