+++ embassy-embedded-hal/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## Unreleased - ReleaseDate
 
+- Shared I2c busses now impl `Clone`
+
 ## 0.5.0 - 2025-08-27
 
 ## 0.4.0 - 2025-08-03